generate all tags in references from ToC (Index)
Programmatically generated the rest of the end-note tags to ensure
consistency and reduce errors. The extra '>' between the tags is
required for Quartz to render correctly. With no space only the
first tag is hidden away at the bottom and the second is left
dangling on its own.

The configuration of Quartz I'm using will automatically populate
the top of each note so I have chosen to hide the tags I want inside
the end-of-note reference.